Color Transform basic done
Former-commit-id: b9cbba63d0ef0c0cbe561129c95b773997e446cc

+++ DIPweek1/affine.cpp
@@ -102,8 +102,8 @@ cv::Mat ResizeTransform(
     int mode
 )
 {
-    const int nRows = scaleX * sourceImg.rows;
-    const int nCols = scaleY * sourceImg.cols;
+    const int nRows = (int)(scaleX * sourceImg.rows);
+    const int nCols = (int)(scaleY * sourceImg.cols);
     double inversescaleX = ((double)sourceImg.rows - 1.0) / (scaleX * (double)sourceImg.rows - 1.0);
     double inversescaleY = ((double)sourceImg.cols - 1.0) / (scaleY * (double)sourceImg.cols - 1.0);
